refactor: rename Trajectories folder to trajectories
