+++ source/state_representation/include/state_representation/trajectories/Trajectory.hpp
